vdk-core: address comments
