+++ projects/vdk-core/src/vdk/internal/plugin/plugin.py
@@ -12,7 +12,7 @@ from vdk.api.plugin.plugin_registry import PluginHookRelay
 from vdk.internal.core.errors import ErrorMessage
 from vdk.internal.core.errors import ResolvableBy
 from vdk.internal.plugin.plugin_manager import VdkPluginManager
-from vdk.internal.util.utils import exit_with_error
+from vdk.internal.util.utils import log_plugin_load_fail
 
 
 log = logging.getLogger(__name__)
@@ -54,9 +54,11 @@ class PluginRegistry(IPluginRegistry):
         try:
             self.__plugin_manager.load_setuptools_entrypoints(self.__group_name)
         except ImportError as e:
-            exit_with_error(ResolvableBy.USER_ERROR, log, e, self.__group_name)
+            log_plugin_load_fail(ResolvableBy.USER_ERROR, log, e, self.__group_name)
+            raise e
         except Exception as e:
-            exit_with_error(ResolvableBy.PLATFORM_ERROR, log, e, self.__group_name)
+            log_plugin_load_fail(ResolvableBy.PLATFORM_ERROR, log, e, self.__group_name)
+            raise e
 
         plugins = self.__plugin_manager.list_name_plugin()
         log.info(
