+++ projects/vdk-core/src/vdk/internal/util/utils.py
@@ -1,23 +1,13 @@
 # Copyright 2021 VMware, Inc.
 # SPDX-License-Identifier: Apache-2.0
 import inspect
-import os
 from logging import Logger
 from typing import Any
 from typing import List
 from typing import Optional
 
-from vdk.api.plugin.plugin_registry import PluginException
-from vdk.internal.builtin_plugins.config.vdk_config import LOG_CONFIG
-from vdk.internal.builtin_plugins.termination_message.writer import (
-    TerminationMessageWriterPlugin,
-)
-from vdk.internal.builtin_plugins.termination_message.writer_configuration import (
-    add_definitions,
-)
+from vdk.internal.core import errors
 from vdk.internal.core.config import Configuration
-from vdk.internal.core.config import ConfigurationBuilder
-from vdk.internal.core.errors import ErrorMessage
 from vdk.internal.core.errors import ResolvableBy
 
 
@@ -55,44 +45,25 @@ def parse_config_sequence(
     return sequence if sequence else []
 
 
-def exit_with_error(
+def log_plugin_load_fail(
     user_error: ResolvableBy, log: Logger, exception: Exception, group_name
 ):
     """
-    Write a termination message, log and exit with specified error.
-    Intended for use in cases when hooks and configuration
-    haven't been initialized yet but we want to write a specific
-    termination message. This scenario occurs when the vdk_main
-    hook hasn't completed yet.
+    Logs errors during plugin load. Calls the errors.log_exception() method
+    which will also update the resolvable context of the VDK run.
+
     :param user_error: If this is a platform error or no
     :param log: The logger which will log the exception
     :param exception: The exception
     :param group_name
     :return:
     """
-    message = ErrorMessage(
-        summary="Plugin load failed",
-        what=f"Cannot load plugin from setuptools entrypoint for group {group_name}",
-        why="See exception for possible reason",
-        consequences="The CLI tool will likely abort.",
-        countermeasures="Re-try again. Check exception message and possibly uninstall a bad "
-        "plugin (pip uninstall) "
-        "Or see what plugins are installed (use `pip list` command) and if "
-        "there aren't issues. "
-        "Or try to reinstall the app in a new clean environment."
-        "Try to revert to previous version of the CLI tool."
-        "If nothing works open a SRE ticket ",
+    errors.log_exception(
+        user_error,
+        log,
+        f"Cannot load plugin from setuptools entrypoint for group {group_name}",
+        "See exception for possible reason",
+        "The CLI tool will likely abort.",
+        "Re-try again. Check exception message and possibly uninstall a bad plugin (pip uninstall) Or see what plugins are installed (use `pip list` command) and if there aren't issues. Or try to reinstall the app in a new clean environment. Try to revert to previous version of the CLI tool. If nothing works open a SRE ticket.",
+        exception,
     )
-    log.error(PluginException(message), exc_info=exception)
-    configuration_builder = ConfigurationBuilder()
-    add_definitions(configuration_builder)
-    configuration_builder.add(key=LOG_CONFIG, default_value="LOCAL")
-
-    configuration = configuration_builder.build()
-
-    writer = TerminationMessageWriterPlugin()
-    user_error_overall = user_error == ResolvableBy.USER_ERROR
-    writer.write_termination_message(
-        error_overall=True, user_error=user_error_overall, configuration=configuration
-    )
-    os._exit(0)
